Add a systemd unit
Add a systemd unit, but do not yet add all of the necessary
macros. They're much improved in Fedora 18, and it's currently unclear
if we want to support F16 and F17, or just start at F18.  The included
unit isn't useful in EPEL, so the legacy initscript should be made
available for those versions.